Rename internal/eval to internal/expr
