+++ internal/expr/expr.go
@@ -1,4 +1,4 @@
-package eval
+package expr
 
 import (
 	"fmt"
