+++ internal/expr/expr_test.go
@@ -1,4 +1,4 @@
-package eval
+package expr
 
 import (
 	"fmt"
